Decouple model from components
