+++ src/model/ProteinNames.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import NameView from './NameView';
-import idx from 'idx';
 import { InfoList } from 'franklin-sites';
 import { ValueWihEvidence } from './types/modelTypes';
 
@@ -28,7 +27,9 @@ export type ProteinNamesData = {
 };
 
 type ProteinNamesDataProps = {
-  data: ProteinNamesData;
+  recommendedName: string;
+  shortNames?: string;
+  alternativeNames?: string[];
 };
 
 const processShortNames = (
@@ -40,49 +41,11 @@ const processShortNames = (
   return shortNameList.reduce((r, d) => `, ${d.value}`, '');
 };
 
-export const processProteinData = (data: ProteinNamesData) => {
-  const recommendedName = idx(
-    data,
-    _ => _.proteinDescription.recommendedName.fullName.value
-  );
-  const alternativeNames: string[] = [];
-  const ecNumbers = idx(
-    data,
-    _ => _.proteinDescription.recommendedName.ecNumbers
-  );
-  if (ecNumbers && ecNumbers.length > 0) {
-    alternativeNames.push(...ecNumbers.map(ec => ec.value));
-  }
-  const submissionNames = idx(data, _ => _.proteinDescription.submissionNames);
-  if (submissionNames && submissionNames.length > 0) {
-    alternativeNames.push(
-      ...submissionNames.map(submissionName => submissionName.fullName.value)
-    );
-  }
-  const alternativeNameArray = idx(
-    data,
-    _ => _.proteinDescription.alternativeNames
-  );
-  if (alternativeNameArray) {
-    const alternativeName = alternativeNameArray.map(
-      altName => altName.fullName.value
-    );
-    if (alternativeName.length) {
-      alternativeNames.push(...alternativeName);
-    }
-  }
-  const shortNames = idx(data, _ =>
-    _.proteinDescription.recommendedName.shortNames
-      .map(shortName => shortName.value)
-      .join(', ')
-  );
-  return { recommendedName, shortNames, alternativeNames };
-};
-
-export const ProteinNames: React.FC<ProteinNamesDataProps> = ({ data }) => {
-  const { recommendedName, shortNames, alternativeNames } = processProteinData(
-    data
-  );
+export const ProteinNames: React.FC<ProteinNamesDataProps> = ({
+  recommendedName,
+  shortNames,
+  alternativeNames,
+}) => {
   const props = {
     name: recommendedName,
     shortNames,
@@ -92,11 +55,10 @@ export const ProteinNames: React.FC<ProteinNamesDataProps> = ({ data }) => {
 };
 
 export const EntryProteinNames: React.FC<ProteinNamesDataProps> = ({
-  data,
+  recommendedName,
+  shortNames,
+  alternativeNames,
 }) => {
-  const { recommendedName, shortNames, alternativeNames } = processProteinData(
-    data
-  );
   const infoData = [
     {
       title: 'Recommended name',
